+++ test/test_query.py
@@ -1,5 +1,6 @@
 from liara.nodes import DocumentNode, StaticNode, Node
 import pathlib
+import pytest
 
 
 class MockDocumentNode(DocumentNode):
@@ -45,3 +46,30 @@ def test_query_filter_by_kind():
 
     doc_children = root.select_children().without_node_kinds('static')
     assert len(list(doc_children)) == 1
+
+
+def test_query_sort_by_missing_field():
+    n1 = MockDocumentNode('/a', {})
+
+    root = Node()
+    root.path = pathlib.PurePosixPath('/')
+    root.add_child(n1)
+
+    with pytest.raises(Exception):
+        list(root.select_children().sorted_by_date())
+
+
+def test_query_sort_by():
+    n1 = MockDocumentNode('/a', {'title': 'A'})
+    n2 = MockDocumentNode('/b', {'title': 'B'})
+
+    root = Node()
+    root.path = pathlib.PurePosixPath('/')
+    root.add_child(n2)
+    root.add_child(n1)
+
+    s1 = list(root.select_children().sorted_by_title())
+    s2 = list(root.select_children().sorted_by_title(reverse=True))
+
+    assert s1[0].url == '/a'
+    assert s2[1].url == '/a'